+++ src/main/java/com/soda/request/dto/response/RequestApproveRequest.java
@@ -1,0 +1,12 @@
+package com.soda.request.dto.response;
+
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+public class RequestApproveRequest {
+    private String comment;
+    private Long projectId;
+    private List<ResponseLinkDTO> links;
+}
