+++ src/main/java/com/soda/request/dto/response/RequestRejectRequest.java
@@ -1,12 +1,11 @@
-package com.soda.request.dto;
+package com.soda.request.dto.response;
 
-import com.soda.request.enums.RequestStatus;
 import lombok.Getter;
 
 import java.util.List;
 
 @Getter
-public class RequestApproveRequest {
+public class RequestRejectRequest {
     private String comment;
     private Long projectId;
     private List<ResponseLinkDTO> links;
