+++ src/main/java/com/soda/request/service/ResponseService.java
@@ -7,21 +7,17 @@ import com.soda.member.enums.MemberProjectRole;
 import com.soda.member.enums.MemberRole;
 import com.soda.member.repository.MemberRepository;
 import com.soda.project.error.ProjectErrorCode;
-import com.soda.request.dto.*;
+import com.soda.request.dto.response.*;
 import com.soda.request.entity.Request;
-import com.soda.request.entity.RequestLink;
 import com.soda.request.entity.Response;
 import com.soda.request.entity.ResponseLink;
 import com.soda.request.repository.RequestRepository;
-import com.soda.request.repository.ResponseLinkRepository;
 import com.soda.request.repository.ResponseRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional(readOnly = true)
